refactor(schemas): reorganize pipeline/ into node/edge/node-data subdirs
- Move 20 flat files into node/, edge/, and node-data/ folders
- Remove NodeTypeSchema (z.string() placeholder); PipelineNode.type now uses
  BuiltinNodeTypeSchema directly since all types are built-in
- Remove unused PipelineEdgeData.dataType field
- Update pipeline-engine consumer to use BUILTIN_NODE_TYPE_ENUM
