+++ packages/schemas/src/pipeline/edge/PipelineEdgeDataSchema.ts
@@ -1,7 +1,6 @@
 import { z } from "zod/v4";
 
 export const PipelineEdgeDataSchema = z.object({
-  label: z.string().optional(),
-  dataType: z.string().optional(),
+  label: z.string().default(""),
 });
 export type PipelineEdgeData = z.infer<typeof PipelineEdgeDataSchema>;
