+++ packages/schemas/src/pipeline/edge/PipelineEdgeSchema.ts
@@ -5,8 +5,8 @@ export const PipelineEdgeSchema = z.object({
   id: z.string(),
   source: z.string(),
   target: z.string(),
-  sourceHandle: z.string().nullable().optional(),
-  targetHandle: z.string().nullable().optional(),
+  sourceHandle: z.string().optional(),
+  targetHandle: z.string().optional(),
   data: PipelineEdgeDataSchema.optional(),
 });
 export type PipelineEdge = z.infer<typeof PipelineEdgeSchema>;
