+++ packages/schemas/src/pipeline/node-data/CompoundNodeDataSchema.ts
