+++ packages/schemas/src/pipeline/node-data/FileObjectNodeDataSchema.ts
