+++ packages/schemas/src/pipeline/node-data/FolderObjectNodeDataSchema.ts
