+++ packages/schemas/src/pipeline/node-data/LocalPathOutputNodeDataSchema.ts
