+++ packages/schemas/src/pipeline/node-data/PipelineNodeDataSchema.ts
