+++ packages/schemas/src/pipeline/node-data/ProjectPathOutputNodeDataSchema.ts
