+++ packages/schemas/src/pipeline/node-data/PromptObjectNodeDataSchema.ts
